+++ libflv/source/mpeg4-hevc.c
@@ -140,7 +140,7 @@ int mpeg4_hevc_decoder_configuration_record_load(const uint8_t* data, size_t byt
 		}
 	}
 
-	hevc->off = dst - hevc->data;
+	hevc->off = (int)(dst - hevc->data);
 	return (int)(p - data);
 }
